Module Common.Naming

Copyright 2024, Artem-Rzhankoff, ItIsMrLag

SPDX-License-Identifier: LGPL-3.0-or-later

val forbidden_names : string list

The user can use these names, but they will be renamed later at the alpha conversion stage.

val alpha_prefix : string

val me_prefix : string

val cc_prefix : string

val ll_prefix : string

val runtime_prefix : string

val anf_prefix : string

val with_pref : string -> string -> string

val reserved_prefs : string list

Reserved prefixes: The user can use them, but they will be renamed during the alpha conversion stage.

INTERNAL LIB

val rt_'main' : string

entry point -- just a name